refactor: 게임 결과와 게임 매니저의 역활 합치도록 리팩토링
